Splits lotus/message module into submodules (#42)
